pyproject.toml
[project]
name = "cat-backend"
package-mode = false
version = "0.0.0"
description = ""
# readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
